Moved hero identification to DotaTools.py
